+++ controlers/updateContactById.js
@@ -1,0 +1,11 @@
+const { Contact } = require("../models");
+const { FindByIdError, ctrlWrap } = require("../helpers");
+const updateContactById = async (req, res, next) => {
+  const { id } = req.params;
+  const result = await Contact.findByIdAndUpdate(id, req.body, {
+    new: true,
+  });
+  FindByIdError(result);
+  res.status(200).json(result);
+};
+module.exports = { updateContactById: ctrlWrap(updateContactById) };
