+++ helpers/ctrlWrap.js
@@ -1,6 +1,7 @@
 const ctrlWrap = (ctrl) => {
   const func = async (req, res, next) => {
     try {
+      console.log("ctrl :>> ", ctrl);
       await ctrl(req, res, next);
     } catch (error) {
       next(error);
