+++ schemas/index.js
@@ -1,3 +1,11 @@
-const { schemas } = require("./schemas");
+const { contactSchema, dataRegexp } = require("./contactSchema");
+const { favoriteSchema } = require("./favoriteSchema");
+const { contactPutSchema } = require("./contactPutSchema");
 
+const schemas = {
+  contactSchema,
+  favoriteSchema,
+  contactPutSchema,
+  dataRegexp,
+};
 module.exports = { schemas };
